Move sideproxy.py to reltools
